+++ CHANGELOG.md
@@ -434,6 +434,9 @@ Miscellaneous:
 
 - 3D viewer: the volume list is now also updated if a volume is deleted.
 
+- 3D viewer: the landmark group list is now updated when landmark groups are
+  added or deleted.
+
 - Volume lists (drop down menus) are now automatically refreshed if a volume is
   added, updated or removed.